+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -7,10 +7,10 @@ import java.math.BigDecimal;
 //*kanae/jaron
 public interface AccountDao {
     BigDecimal getBalance(String user);
-    int findIdByAccountId(int accountId);
+    int findIdByAccountId(long accountId);
 
-    Account getAccountByUserId(int userId);
-    Account getAccountByAccountId(int accountId);
+    Account getAccountByUserId(long userId);
+    Account getAccountByAccountId(long accountId);
 
     void updateAccount(Account account);
